+++ sdk/python/kfp/dsl/__init__.py
@@ -15,8 +15,7 @@
 
 from ._pipeline_param import PipelineParam, match_serialized_pipelineparam
 from ._pipeline import Pipeline, pipeline, get_pipeline_conf
-from ._base_op import Sidecar
-from ._container_op import ContainerOp
+from ._container_op import ContainerOp, Sidecar
 from ._resource_op import ResourceOp
 from ._volume_op import (
     VolumeOp, VOLUME_MODE_RWO, VOLUME_MODE_RWM, VOLUME_MODE_ROM
